[Frontend] Extract shared PageReader component from CBZReader
